Read-only-repositories can be subscribed too

+++ gim_project/subscriptions/models.py
@@ -4,6 +4,7 @@ from extended_choices import Choices
 
 from core.models import GithubUser, Repository
 
+
 WAITING_SUBSCRIPTION_STATES = Choices(
     ('WAITING', 1, 'Waiting'),
     ('FETCHING', 2, 'Fetching'),
@@ -16,7 +17,6 @@ class WaitingSubscription(models.Model):
     repository_name = models.TextField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
-    is_admin = models.BooleanField(default=False)
     state = models.PositiveSmallIntegerField(
         choices=WAITING_SUBSCRIPTION_STATES.CHOICES,
         default=WAITING_SUBSCRIPTION_STATES.WAITING)
@@ -41,29 +41,36 @@ class WaitingSubscription(models.Model):
             name=repository_name
         )
 
-    def convert(self):
+    def convert(self, rights):
         """
-        Convert the waiting subscription into a real one.
+        Convert the waiting subscription into a real one with the given rights
+        (rights are "admin", "push", "read")
         The repository matching he repository_name field must exist.
         """
         if self.state in (WAITING_SUBSCRIPTION_STATES.FAILED,):
             raise Exception('Cannot convert a failed waiting subscription')
+
+        if rights not in set(["admin", "push", "read"]):
+            raise Exception('Cannot convert a subscription if not enough rights')
 
         owner_username, repository_name = self.repository_name.split('/')
         subscription, _ = Subscription.objects.get_or_create(
             user=self.user,
             repository=self.repository
         )
-        subscription.state = SUBSCRIPTION_STATES.ADMIN if self.is_admin else SUBSCRIPTION_STATES.SIMPLE
+        subscription.state = SUBSCRIPTION_STATES.ADMIN if rights == "admin" \
+                             else SUBSCRIPTION_STATES.USER if rights == "push" \
+                             else SUBSCRIPTION_STATES.READ
         subscription.save()
 
         self.delete()
 
 
 SUBSCRIPTION_STATES = Choices(
-    ('SIMPLE', 1, 'User'),
-    ('ADMIN', 2, 'Admin'),
-    ('NORIGHTS', 3, 'No rights'),
+    ('READ', 1, 'Read-only'),  # can read
+    ('USER', 2, 'User'),  # can push, create issues
+    ('ADMIN', 3, 'Admin'),  # can admin, push, create issues
+    ('NORIGHTS', 4, 'No rights'),  # no access
 )
 
 
@@ -74,7 +81,7 @@ class Subscription(models.Model):
     updated_at = models.DateTimeField(auto_now=True)
     state = models.PositiveSmallIntegerField(
         choices=SUBSCRIPTION_STATES.CHOICES,
-        default=SUBSCRIPTION_STATES.SIMPLE)
+        default=SUBSCRIPTION_STATES.READ)
 
     class Meta:
         unique_together = [('user', 'repository'), ]
